# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "etare-platform API",
    "version": "v1",
    "description": "API métier de la plateforme ETARE numérique. Toutes les réponses sont limitées au SIS actif et aux droits de l’appelant."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Jeton d’accès du fournisseur d’identité (Supabase Auth au MVP)."
      }
    },
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "UNAUTHENTICATED",
                  "FORBIDDEN",
                  "MFA_REQUIRED",
                  "TENANT_REQUIRED",
                  "NOT_FOUND",
                  "VALIDATION_FAILED",
                  "CONFLICT",
                  "PRECONDITION_FAILED",
                  "PRECONDITION_REQUIRED",
                  "PAYLOAD_TOO_LARGE",
                  "SERVICE_UNAVAILABLE",
                  "INVALID_TRANSITION",
                  "SELF_APPROVAL_FORBIDDEN",
                  "RATE_LIMITED",
                  "INTERNAL"
                ]
              },
              "message": {
                "type": "string"
              },
              "trace_id": {
                "type": "string"
              },
              "fields": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "path": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "path",
                    "message"
                  ]
                }
              }
            },
            "required": [
              "code",
              "message",
              "trace_id"
            ]
          }
        },
        "required": [
          "error"
        ],
        "description": "Erreur structurée : code stable, message utile, trace_id, sans détail interne."
      },
...
      "SiteList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SiteSummary"
            }
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ],
            "minLength": 1,
            "maxLength": 512,
            "pattern": "^[A-Za-z0-9_-]+$"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ]
      },
...
      "LevelCreate": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "minLength": 1,
            "maxLength": 40
          },
          "sort_order": {
            "type": "integer",
            "minimum": -20,
            "maximum": 200
          },
          "elevation_m": {
            "type": [
              "number",
              "null"
            ],
            "minimum": -200,
            "maximum": 1000
          }
        },
        "required": [
          "label",
          "sort_order"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/sites": {
      "get": {
        "operationId": "listSites",
        "summary": "Rechercher les sites du SIS actif (données de travail)",
        "tags": [
          "sites"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 25
            },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 512,
              "pattern": "^[A-Za-z0-9_-]+$"
            },
            "required": false,
            "name": "cursor",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 2,
              "maxLength": 100
            },
            "required": false,
            "name": "q",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "erp",
                "industrial",
                "health",
                "education",
                "heritage",
                "other"
              ]
            },
            "required": false,
            "name": "site_type",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "draft",
                "active",
                "inactive",
                "archived"
              ]
            },
            "required": false,
            "name": "status",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 120
            },
            "required": false,
            "name": "city",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "SIS actif. Indication seulement : l’appartenance est revérifiée côté serveur."
            },
            "required": true,
            "description": "SIS actif. Indication seulement : l’appartenance est revérifiée côté serveur.",
            "name": "x-tenant-id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteList"
                }
              }
            }
          },
          "400": {
            "description": "Requête invalide (VALIDATION_FAILED, TENANT_REQUIRED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Authentification requise",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Accès refusé (FORBIDDEN, MFA_REQUIRED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Erreur interne (sans détail)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createSite",
        "summary": "Créer un site",
        "tags": [
          "sites"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "SIS actif. Indication seulement : l’appartenance est revérifiée côté serveur."
            },
            "required": true,
            "description": "SIS actif. Indication seulement : l’appartenance est revérifiée côté serveur.",
            "name": "x-tenant-id",
            "in": "header"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SiteCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Succès",
            "headers": {
              "ETag": {
                "description": "Version de la ressource (row_version)",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteDetail"
                }
              }
            }
          },
          "400": {
            "description": "Requête invalide (VALIDATION_FAILED, TENANT_REQUIRED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Authentification requise",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Accès refusé (FORBIDDEN, MFA_REQUIRED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflit avec des données existantes (CONFLICT)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Erreur interne (sans détail)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/sites/{id}": {
      "get": {
        "operationId": "getSite",
        "summary": "Détail d’un site du SIS actif",
        "tags": [
          "sites"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "SIS actif. Indication seulement : l’appartenance est revérifiée côté serveur."
            },
            "required": true,
            "description": "SIS actif. Indication seulement : l’appartenance est revérifiée côté serveur.",
            "name": "x-tenant-id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "headers": {
              "ETag": {
                "description": "Version de la ressource (row_version)",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteDetail"
                }
              }
            }
          },
          "400": {
            "description": "Requête invalide (VALIDATION_FAILED, TENANT_REQUIRED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Authentification requise",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Accès refusé (FORBIDDEN, MFA_REQUIRED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Ressource introuvable ou non autorisée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Erreur interne (sans détail)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "updateSite",
        "summary": "Modifier un site",
        "tags": [
          "sites"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "SIS actif. Indication seulement : l’appartenance est revérifiée côté serveur."
            },
            "required": true,
            "description": "SIS actif. Indication seulement : l’appartenance est revérifiée côté serveur.",
            "name": "x-tenant-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Version sur laquelle porte la modification (valeur de l’ETag reçu), ex. \"3\".",
              "example": "\"3\""
            },
            "required": true,
            "description": "Version sur laquelle porte la modification (valeur de l’ETag reçu), ex. \"3\".",
            "name": "if-match",
            "in": "header"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SiteUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Succès",
            "headers": {
              "ETag": {
                "description": "Version de la ressource (row_version)",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteDetail"
                }
              }
            }
          },
          "400": {
            "description": "Requête invalide (VALIDATION_FAILED, TENANT_REQUIRED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Authentification requise",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Accès refusé (FORBIDDEN, MFA_REQUIRED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Ressource introuvable ou non autorisée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflit avec des données existantes (CONFLICT)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "412": {
            "description": "Modifiée entre-temps : relire puis réessayer (PRECONDITION_FAILED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "428": {
            "description": "En-tête If-Match manquant (PRECONDITION_REQUIRED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Erreur interne (sans détail)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/buildings/{id}/levels": {
      "post": {
        "operationId": "createLevel",
        "summary": "Ajouter un niveau à un bâtiment",
        "tags": [
          "buildings"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "SIS actif. Indication seulement : l’appartenance est revérifiée côté serveur."
            },
            "required": true,
            "description": "SIS actif. Indication seulement : l’appartenance est revérifiée côté serveur.",
            "name": "x-tenant-id",
            "in": "header"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LevelCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Succès",
            "headers": {
              "ETag": {
                "description": "Version de la ressource (row_version)",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Level"
                }
              }
            }
          },
          "400": {
            "description": "Requête invalide (VALIDATION_FAILED, TENANT_REQUIRED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Authentification requise",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Accès refusé (FORBIDDEN, MFA_REQUIRED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Ressource introuvable ou non autorisée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflit avec des données existantes (CONFLICT)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Erreur interne (sans détail)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}